pyproject.toml
[project]
name = "multipurpose-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Shishupal Rajpurohit",email = "shishupalrajpurohit2000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "chainlit (>=2.8.3,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-core (>=0.3.78,<0.4.0)",
    "langchain-groq (>=0.3.8,<0.4.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "pypdf (>=6.1.1,<7.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "tavily-python (>=0.7.12,<0.8.0)",
    "sentence-transformers (>=5.1.1,<6.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "cryptography (>=46.0.2,<47.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
